pyproject.toml
[project]
name = "autonomous-financial-workstation-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0.post1",
    "firecrawl-py>=2.16.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.8",
    "langchain-openai>=0.3.28",
    "langchain-tavily>=0.2.9",
    "langgraph>=0.5.3",
    "python-dotenv>=1.1.1",
]
